refactor: This is a simple react component 🤓
