+++ src/ducks/onboarding/Onboarding.jsx
@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React from 'react'
 import { translate, Button, withBreakpoints } from 'cozy-ui/react'
 import { flowRight as compose } from 'lodash'
 
@@ -21,53 +21,49 @@ import {
   CTA
 } from 'cozy-ui/react/Hero'
 
-class Onboarding extends Component {
-  render() {
-    const {
-      t,
-      breakpoints: { isMobile }
-    } = this.props
+const Onboarding = props => {
+  const {
+    t,
+    breakpoints: { isMobile }
+  } = props
 
-    return (
-      <Hero>
-        <Topbar>
-          <PageTitle>
-            {t(`Onboarding.title.${isMobile ? 'mobile' : 'desktop'}`)}
-          </PageTitle>
-        </Topbar>
-        <Sections>
+  return (
+    <Hero>
+      <Topbar>
+        <PageTitle>
+          {t(`Onboarding.title.${isMobile ? 'mobile' : 'desktop'}`)}
+        </PageTitle>
+      </Topbar>
+      <Sections>
+        <Section>
+          <Icon color={palette.pomegranate} icon={calculator} />
+          <Subtitle>{t('Onboarding.manage-budget.title')}</Subtitle>
+          <Paragraph>{t('Onboarding.manage-budget.description')}</Paragraph>
+        </Section>
+        {!isMobile ? (
           <Section>
-            <Icon color={palette.pomegranate} icon={calculator} />
-            <Subtitle>{t('Onboarding.manage-budget.title')}</Subtitle>
-            <Paragraph>{t('Onboarding.manage-budget.description')}</Paragraph>
+            <Icon color={palette.portage} icon={watch} />
+            <Subtitle>{t('Onboarding.save-time.title')}</Subtitle>
+            <Paragraph>{t('Onboarding.save-time.description')}</Paragraph>
           </Section>
-          {!isMobile ? (
-            <Section>
-              <Icon color={palette.portage} icon={watch} />
-              <Subtitle>{t('Onboarding.save-time.title')}</Subtitle>
-              <Paragraph>{t('Onboarding.save-time.description')}</Paragraph>
-            </Section>
-          ) : null}
-          {!isMobile ? (
-            <Section>
-              <Icon color={palette['dodgerBlue']} icon={cozy} />
-              <Subtitle>{t('Onboarding.cozy-assistant.title')}</Subtitle>
-              <Paragraph>
-                {t('Onboarding.cozy-assistant.description')}
-              </Paragraph>
-            </Section>
-          ) : null}
-        </Sections>
-        <CTA>
-          <CollectLink>
-            <Button theme="regular">
-              {t('Onboarding.connect-bank-account')}
-            </Button>
-          </CollectLink>
-        </CTA>
-      </Hero>
-    )
-  }
+        ) : null}
+        {!isMobile ? (
+          <Section>
+            <Icon color={palette['dodgerBlue']} icon={cozy} />
+            <Subtitle>{t('Onboarding.cozy-assistant.title')}</Subtitle>
+            <Paragraph>{t('Onboarding.cozy-assistant.description')}</Paragraph>
+          </Section>
+        ) : null}
+      </Sections>
+      <CTA>
+        <CollectLink>
+          <Button theme="regular">
+            {t('Onboarding.connect-bank-account')}
+          </Button>
+        </CollectLink>
+      </CTA>
+    </Hero>
+  )
 }
 
 export default compose(translate(), withBreakpoints())(Onboarding)
