+++ src/ducks/onboarding/index.js
@@ -1,5 +1,5 @@
 import Onboarding from './Onboarding'
-import EnsureHasAccounts from './ensureHasAccounts'
+import EnsureHasAccounts from './EnsureHasAccounts'
 import EnsureIsFirstSynced from './ensureIsFirstSynced'
 
 export { Onboarding, EnsureIsFirstSynced, EnsureHasAccounts }
